Squashed commit of the following:
commit 1daf8d73171e9496694d32cfce1113d0bb0f5019
Author: Locke Huang <locke.huang@gmail.com>
Date:   Thu Nov 13 07:23:02 2025 +1300

    fix warning

commit e1ee2a74b0504d2fe1c2fe2678b9c67046d4172b
Author: Locke Huang <locke.huang@gmail.com>
Date:   Thu Nov 13 07:20:17 2025 +1300

    fix windows key issue.

commit ea1065e71f285211a2648378b1b7bf69ddf5f4db
Author: Locke Huang <locke.huang@gmail.com>
Date:   Thu Nov 13 06:53:15 2025 +1300

    fix windows key press issue(with pathes). next commit will have compelet change.

commit 44050dd47179a202a3f1ffee3aa9c05b9d59127e
Author: Locke Huang <locke.huang@gmail.com>
Date:   Thu Nov 13 05:33:02 2025 +1300

    fix server/client setting path issue for all platforms

commit d0536284bc3e0d148cc57bf3e9773f93486d876d
Author: Locke Huang <locke.huang@gmail.com>
Date:   Thu Nov 13 05:19:28 2025 +1300

    codex review code change.

commit 164fbf0ed149362a10c746740d5d273731119744
Author: Locke Huang <locke.huang@gmail.com>
Date:   Thu Nov 13 04:54:28 2025 +1300

    fix the mouse hook issue on windows

commit d6f15f87009480383b76986db86d1e50cce0d970
Author: Locke Huang <locke.huang@gmail.com>
Date:   Wed Nov 12 20:34:35 2025 +1300

    TLS now can connect.

commit e38106905244d85f89d345d0111867ed56c2ad35
Author: Locke Huang <locke.huang@gmail.com>
Date:   Wed Nov 12 20:08:54 2025 +1300

    non-tls connection seems working, but server does not have enter zone.

+++ src/lib/platform/bridge/CdcTransport.cpp
@@ -175,7 +175,7 @@ bool CdcTransport::open()
     return false;
   }
 
-  m_fd = reinterpret_cast<int>(handle);
+  m_fd = reinterpret_cast<intptr_t>(handle);
 
   DCB dcb = {0};
   dcb.DCBlength = sizeof(DCB);
